[Design] 실시간 경기 분석 데이터 컴포넌트화 #1
