+++ src/app/page.tsx
@@ -1,10 +1,10 @@
-import WeekSchedule from "@/components/WeekSchedule/WeekSchedule";
-import LeagueRanking from "@/components/LeagueRanking/LeagueRanking";
-import Notice from "@/components/Notice/Notice";
-import GameInfo from "@/components/GameInfo/GameInfo";
-import Shortcuts from "@/components/Shortcuts/Shortcuts";
-import MatchBet from "@/components/MatchBet/MatchBet";
-import Sponsor from "@/components/Sponsor/Sponsor";
+import WeekSchedule from "@/components/MainPage/WeekSchedule/WeekSchedule";
+import LeagueRanking from "@/components/MainPage/LeagueRanking/LeagueRanking";
+import Notice from "@/components/MainPage/Notice/Notice";
+import GameInfo from "@/components/MainPage/GameInfo/GameInfo";
+import Shortcuts from "@/components/MainPage/Shortcuts/Shortcuts";
+import MatchBet from "@/components/MainPage/MatchBet/MatchBet";
+import Sponsor from "@/components/MainPage/Sponsor/Sponsor";
 import Chatting from "@/components/Chatting/Chatting";
 import Floating from "@/components/Floating/Floating";
 import ChatSocketConnectHandler from "@/components/Socket/ChatSocketConnectHandler";
