+++ src/components/MainPage/GameInfo/Lineup.tsx
