+++ src/components/MainPage/GameInfo/Match/Match.tsx
