+++ src/components/MainPage/LeagueRanking/RankingList.tsx
