+++ src/components/MainPage/Notice/Notice.tsx
