+++ src/components/MainPage/Shortcuts/Shortcuts.tsx
