+++ src/components/MainPage/Sponsor/CustomLeftArrow.tsx
