+++ src/components/MainPage/Sponsor/CustomRightArrow.tsx
