+++ src/components/MainPage/Sponsor/Sponsor.tsx
