+++ src/components/MainPage/WeekSchedule/WeekSchedule.tsx
