+++ src/types/gameInfo.ts
@@ -1,0 +1,11 @@
+export type TselectRecord = "CURRENT" | "RELATIVE";
+
+export type TUseSelectPitcherRecord = {
+  selectPitcherRecord: TselectRecord;
+  setSelectPitcherRecord: (v: TselectRecord) => void;
+};
+
+export type TUseSelectHitterRecord = {
+  selectHitterRecord: TselectRecord;
+  setSelectHitterRecord: (v: TselectRecord) => void;
+};
